refactor(core): implement skill-based architecture, Event protocol & deep self-healing
• Skills & Context Optimization: Refactored monolithic agents into granular skills (SystemInfo, AudioControl, etc.). The orchestrator now dynamically queries '/tools/list' with skill-filtering to inject only the necessary tool schemas, preventing token bloat and context memory overhead.

• Event Protocol Migration: Replaced bloated chunk payload structures with a unified, lightweight 'Event' schema (Start, Thinking, Answer, Error, Finish) to streamline SSE serialization and reduce IPC packet latency over local Unix Domain Sockets.

• Deep Self-Healing Loop: Implemented robust recovery for model anomalies (such as empty token generation, hallucinated tools, or corrupted downstream JSON payloads). The loop now converts failures into contextual system prompts and performs recursive evaluation up to configured 'max_retries' limit, with explicit critical-path bypass.

• Endpoints Update: Migrated agent startup validation to the fast '/init' hook, replaced the hot-reload endpoint with '/refresh', and trimmed down agent metadata storage to zero out idle RAM footprint.

+++ crates/share/src/session_id.rs
@@ -10,7 +10,7 @@ use std::{num::ParseIntError, str::FromStr};
 pub struct SessionId {
     pub user_id: u128,
     pub timestamp: u128,
-    pub salt: u8,
+    pub salt: u16,
 }
 
 impl SessionId {
@@ -66,7 +66,7 @@ impl FromStr for SessionId {
             .parse::<u128>()
             .map_err(SessionIdError::InvalidTimestamp)?;
         let salt = salt_str
-            .parse::<u8>()
+            .parse::<u16>()
             .map_err(SessionIdError::InvalidSalt)?;
 
         Ok(Self {
